change: finalize refactor
