+++ server/rosetta.go
@@ -1,13 +1,10 @@
 package server
 
 import (
-	"github.com/spf13/cobra"
-	"github.com/tendermint/cosmos-rosetta-gateway/service"
-
 	"github.com/cosmos/cosmos-sdk/codec"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
-	"github.com/cosmos/cosmos-sdk/server/rosetta"
-	"github.com/cosmos/cosmos-sdk/server/rosetta/config"
+	rosettacfg "github.com/cosmos/cosmos-sdk/server/rosetta/config"
+	"github.com/spf13/cobra"
 )
 
 // RosettaCommand builds the rosetta root command given
@@ -17,7 +14,7 @@ func RosettaCommand(ir codectypes.InterfaceRegistry, cdc codec.Marshaler) *cobra
 		Use:   "rosetta",
 		Short: "spin up a rosetta server",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			conf, err := config.FromFlags(cmd.Flags())
+			conf, err := rosettacfg.FromFlags(cmd.Flags())
 			if err != nil {
 				return err
 			}
@@ -25,26 +22,14 @@ func RosettaCommand(ir codectypes.InterfaceRegistry, cdc codec.Marshaler) *cobra
 			if protoCodec, ok := cdc.(*codec.ProtoCodec); ok {
 				conf.WithCodec(ir, protoCodec)
 			}
-			if err := conf.Validate(); err != nil {
-				return err
-			}
-
-			adapter, client, err := config.RetryRosettaFromConfig(conf)
+			rosettaSrv, err := rosettacfg.HandlerFromConfig(conf)
 			if err != nil {
 				return err
 			}
-
-			svc, err := service.New(
-				service.Options{ListenAddress: conf.Addr},
-				rosetta.NewNetwork(conf.NetworkIdentifier(), adapter, client),
-			)
-			if err != nil {
-				return err
-			}
-			return svc.Start()
+			return rosettaSrv.Start()
 		},
 	}
-	config.SetFlags(cmd.Flags())
+	rosettacfg.SetFlags(cmd.Flags())
 
 	return cmd
 }
