+++ server/start.go
@@ -4,6 +4,7 @@ package server
 
 import (
 	"fmt"
+	"github.com/cosmos/cosmos-sdk/codec"
 	"os"
 	"runtime/pprof"
 	"time"
@@ -11,7 +12,7 @@ import (
 	"github.com/spf13/cobra"
 	"google.golang.org/grpc"
 
-	"github.com/tendermint/cosmos-rosetta-gateway/service"
+	crgserver "github.com/tendermint/cosmos-rosetta-gateway/server"
 	"github.com/tendermint/tendermint/abci/server"
 	tcmd "github.com/tendermint/tendermint/cmd/tendermint/commands"
 	tmos "github.com/tendermint/tendermint/libs/os"
@@ -26,7 +27,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/server/api"
 	"github.com/cosmos/cosmos-sdk/server/config"
 	servergrpc "github.com/cosmos/cosmos-sdk/server/grpc"
-	"github.com/cosmos/cosmos-sdk/server/rosetta"
 	rosettacfg "github.com/cosmos/cosmos-sdk/server/rosetta/config"
 	"github.com/cosmos/cosmos-sdk/server/types"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
@@ -313,7 +313,7 @@ func startInProcess(ctx *Context, clientCtx client.Context, appCreator types.App
 		}
 	}
 
-	var rosettaSrv *service.Service
+	var rosettaSrv crgserver.Handler
 	if config.Rosetta.Enable {
 		offlineMode := config.Rosetta.Offline
 		if !config.GRPC.Enable { // If GRPC is not enabled rosetta cannot work in online mode, so it works in offline mode.
@@ -329,25 +329,12 @@ func startInProcess(ctx *Context, clientCtx client.Context, appCreator types.App
 			Retries:       config.Rosetta.Retries,
 			Offline:       offlineMode,
 		}
+		conf.WithCodec(clientCtx.InterfaceRegistry, clientCtx.JSONMarshaler.(*codec.ProtoCodec))
 
-		err = conf.Validate()
+		rosettaSrv, err = rosettacfg.HandlerFromConfig(conf)
 		if err != nil {
 			return err
 		}
-
-		adapter, rosClient, err := rosettacfg.RetryRosettaFromConfig(conf)
-		if err != nil {
-			return err
-		}
-
-		rosettaSrv, err = service.New(
-			service.Options{ListenAddress: conf.Addr},
-			rosetta.NewNetwork(conf.NetworkIdentifier(), adapter, rosClient),
-		)
-		if err != nil {
-			return err
-		}
-
 		errCh := make(chan error)
 		go func() {
 			if err := rosettaSrv.Start(); err != nil {
